migrate the last test and delete e2e dir
