+++ tests/data/cow_protocol.VouchRegister_evt_InvalidateVouch.csv
@@ -1,0 +1,2 @@
+solver,bondingPool,sender,contract_address,evt_tx_hash,evt_index,evt_block_time,evt_block_number
+\x1e05ee95b8dfcb7c7567fa52e0ef47288b1c8fb5,\x5d4020b9261f01b6f8a45db929704b0ad6f5e9e6,\x423cec87f19f0778f549846e0801ee267a917935,\xb422f2520b0b7fd86f7da61b32cc631a59ed7e8f,\x7b96c5e0238dd0c80391a6fd0bbb5ee7be607cdeedd355c4ffb51ddb03f36727,273,2022-05-04T17:29:05+00:00,14712269
